Rename the sample app's data-fetcher to data-fetcher-factory to better describe what it's doing and to make it consistent with our other files
